Updated HomeScreen
Updated home screen so when the user navigates to it it will automatically create the buttons that correspond with classes that the user has already joined. This list first queries the classRelation table to get the classes that the user is in and then uses the courseIds from that query to query the class table to get the rest of the information about the classes.